requirements.txt
accelerate==0.20.3
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
black==23.3.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
datasets==2.13.2
dill==0.3.6
docker-pycreds==0.4.0
exceptiongroup==1.2.2
filelock==3.12.2
flake8==5.0.4
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==4.2.0
iniconfig==2.0.0
mccabe==0.7.0
memory-profiler==0.61.0
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==1.0.0
numpy==1.21.6
packaging==24.0
pandas==1.3.5
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.2.0
protobuf==4.24.4
psutil==6.1.0
pyarrow==12.0.1
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.5
scipy==1.7.3
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tokenizers==0.13.3
tomli==2.0.1
#torch==1.13.1+cpu
torch==1.13.1
tqdm==4.67.0
transformers==4.30.2
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.18.6
xxhash==3.5.0
yarl==1.9.4
zipp==3.15.0
